I want this to be very clear. I hate it Richard coming up with ideas so good that I cannot go home until they are implemented. Especially when implementing the good idea requires me to learn a new JavaScript Library. Really, this is beyond a joke.
Check the email address that the user enters. If it is from GMail, Yahoo! Mail
or Hotmail, show only the link to the help from that system. Do this in JQuery,
without conflicting with the existing Prototype code.

--HG--
extra : convert_revision : svn%3A01e103ed-f575-20b1-d296-6440ccab0075/GSProfile/trunk%402688